Add end game menu
