+++ Pandemonium/Assets/Prefab/Wave/SkeletonHeadActiveWave.prefab
@@ -66,7 +66,7 @@ PrefabInstance:
     m_Modifications:
     - target: {fileID: 6584415696885373002, guid: 524054644971d5a47aedffdff4bd95a4, type: 3}
       propertyPath: m_Name
-      value: P_PROP_skeleton_skull_dungeon_01
+      value: SkeletonHeadActiveWave
       objectReference: {fileID: 0}
     - target: {fileID: 7871116668268522073, guid: 524054644971d5a47aedffdff4bd95a4, type: 3}
       propertyPath: m_RootOrder
@@ -119,3 +119,20 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 7871116668268522073, guid: 524054644971d5a47aedffdff4bd95a4, type: 3}
   m_PrefabInstance: {fileID: 7070551411665560946}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &4143081021198155064 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 6584415696885373002, guid: 524054644971d5a47aedffdff4bd95a4, type: 3}
+  m_PrefabInstance: {fileID: 7070551411665560946}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &7324749166235973460
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4143081021198155064}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f6ec8182851ac4d43be2da0f4eb73f90, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
