+++ Pandemonium/Assets/Scripts/GameManager.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using Unity.Netcode;
 using UnityEngine;
 
@@ -39,29 +37,6 @@ public class GameManager : NetworkBehaviour {
     private void OnDisconnectedFromServer()
     {
         StartSession.instance.CancelSession();
-    }
-
-    public void Defeat()
-    {
-        if (!IsServer && !GameManager.Instance.IsSolo()) return;
-
-        GameObject[] enemies = GameObject.FindGameObjectsWithTag("Enemy");
-
-        foreach (GameObject enemy in enemies)
-        {
-            if (IsServer) enemy.GetComponent<NetworkObject>().Despawn(true);
-            else Destroy(enemy);
-        }
-
-        Debug.Log("Defeat");
-        WaveManager.Instance.StopWave();
-    }
-
-    public void Victory()
-    {
-        if (!IsServer && !GameManager.Instance.IsSolo()) return;
-
-        Debug.Log("Victory");
     }
 
     public GameObject GetSkeletonPrefab(){return skeletonPrefab;}
